pyproject.toml
[tool.poetry]
name = "uchacks"
version = "0.1.0"
description = ""
authors = ["William Fecteau <wfecteau@edu.cegepgarneau.ca>"]

[tool.poetry.dependencies]
python = "^3.9"
"discord.py" = "^1.6.0"
python-dotenv = "^0.15.0"
pyodbc = "^4.0.30"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
